Refactor feedback handling
